Add newline to end of .env file in edit-env
I'm petty so it has to be there.

But nah, really, it's just better for terminal output if you ever need to cat .env. Without the new line, the terminal just keeps going from that line, which doesn't look nice. If you have the new line there, the terminal prompt nicely starts on a new line.