Include git fetch --prune in merge mode of git-post-mergee-cleanup too
The extra branches that should've been deleted are still persisting even in repositories that use merge mode for some reason. As such, I've decided to take it out of just the rebase mode workflow and have it so that it always runs so dead remote branches are always cleaned up.