Add script to check for existence of file dependencies
This is useful when using the new use-local-components and use-live-components script, where we may sometimes forget to go back to using the live version of the package before committing. In cases like that, this script can be added to the linting script to ensure that local dependencies do not make it into production.