Add command to check for version number changes
This is useful to have in a CI script so that we can enforce that changes to source code must also change the version number. This helps to ensure that the most recent published version of a package definitely matches the source code, reducing the change of any discrepancy between them that may make debugging harder.