Add a set-release-status command
This will automatically change the status of a release document to be released so that we don't have to do it as a second step before release. Ideally I would like this to go in the commit-version-change workflow where, if it finds a release document, it changes the package.json version as well as the release document status.