Fix tests setup
First of all, I got the wrong package manager this entire time...

Second of all, I think it's better to just run the build step in the actual test script itself to guarantee that we get a freshly built package, rather than screw around with the setup file. I may've gotten away with it with NPM and/or tsup, but PNPM and tsdown don't like that as much for whatever reason. I think this approach is safer anyway as it guarantees that we only build once, at the very start of the test.